app/requirements.txt
flask==2.3.3
numpy==1.24.3
tensorflow==2.19.0
opencv-python==4.8.0.76
pillow==10.0.0